+++ app/Http/Controllers/ProdukController.php
@@ -6,16 +6,17 @@ use DB, Auth;
 use App\Models\Produk;
 use App\Http\Requests\StoreProdukRequest;
 use App\Http\Requests\UpdateProdukRequest;
+use App\Models\Kategori;
 
 class ProdukController extends Controller
 {
-    // function __construct()
-    // {
-    //      $this->middleware('permission:view_produk', ['only' => ['index','show']]);
-    //      $this->middleware('permission:add_produk', ['only' => ['create','store']]);
-    //      $this->middleware('permission:edit_produk', ['only' => ['edit','update']]);
-    //      $this->middleware('permission:delete_produk', ['only' => ['destroy']]);
-    // }
+    function __construct()
+    {
+        $this->middleware('permission:view_produk', ['only' => ['index', 'show']]);
+        $this->middleware('permission:add_produk', ['only' => ['create', 'store']]);
+        $this->middleware('permission:edit_produk', ['only' => ['edit', 'update']]);
+        $this->middleware('permission:delete_produk', ['only' => ['destroy']]);
+    }
 
     /**
      * Display a listing of the resource.
@@ -35,7 +36,9 @@ class ProdukController extends Controller
     public function create()
     {
         $kategoris = DB::table('kategoris')->where('sekolah_id', Auth::user()->sekolah_id)->get();
-        return view('produk.form', compact('kategoris'));
+        $subcategories = DB::table('subcategories')->get();
+        $jurusans = DB::table('jurusans')->get();
+        return view('produk.form', compact('kategoris', 'subcategories', 'jurusans'));
     }
 
     /**
@@ -46,21 +49,33 @@ class ProdukController extends Controller
      */
     public function store(StoreProdukRequest $request)
     {
+        try {
+            for ($i = 0; $i < $request->jumlah; $i++) {
+                $getCategorieName = DB::table('kategoris')->where('id', $request->kategori_id)->first()->kode;
+                $countProdSub = count(DB::table('produks')->where('kategori_id', $request->kategori_id)->where('sub_kategori_id', $request->sub_kategori_id)->get()) + 1;
+                $countProdKat = count(DB::table('produks')->where('kategori_id', $request->kategori_id)->get()) + 1;
+                // dd("$getCategorieName".count($countProd) + 1);
 
-        // dd($request->all());
-        $produk = Produk::create([
-            'kategori_id' => $request->katgori_id,
-            'sub_kategori_id' => $request->sub_katgori_id,
-            'jurusan_id' => $request->jurusan_id,
-            'nama' => $request->nama,
-            'kode' => $request->kode,
-            'merek' => $request->merek,
-            'kondisi' => $request->kondisi,
-            'ket_produk' => $request->ket_produk,
-            'ket_kondisi' => $request->ket_kondisi,
-        ]); 
+                $produk = Produk::create([
+                    'kategori_id' => $request->kategori_id,
+                    'sub_kategori_id' => $request->sub_kategori_id,
+                    'jurusan_id' => $request->jurusan_id,
+                    'nama' => $request->nama,
+                    'kode' => "$getCategorieName-$countProdKat-$countProdSub",
+                    'merek' => $request->merek,
+                    'kondisi' => $request->kondisi,
+                    'ket_produk' => $request->ket_produk,
+                    'ket_kondisi' => $request->ket_kondisi,
+                ]);
 
-        return redirect('/produk')->with('success', 'Produk berhasil ddi buat');
+                insertLog(Auth::user()->name . " Berhasil menambahkan produk " . $produk['nama']);
+            }
+
+            return redirect()->route('produk.index')->with('msg_success', 'Berhasil menambahkan produk');
+        } catch (\Throwable $th) {
+            dd($th);
+            return redirect()->route('produk.index')->with('msg_error', 'Gagal Ditambahkan');
+        }
     }
 
     /**
@@ -71,7 +86,7 @@ class ProdukController extends Controller
      */
     public function show(Produk $produk)
     {
-        //
+        return $produk;
     }
 
     /**
@@ -82,7 +97,7 @@ class ProdukController extends Controller
      */
     public function edit(Produk $produk)
     {
-        //
+        return view('produk.form', compact('produk'));
     }
 
     /**
@@ -94,17 +109,29 @@ class ProdukController extends Controller
      */
     public function update(UpdateProdukRequest $request, Produk $produk)
     {
-        $produk->update([
-            'kategori_id' => $request->katgori_id,
-            'sub_kategori_id' => $request->sub_katgori_id,
-            'jurusan_id' => $request->jurusan_id,
-            'nama' => $request->nama,
-            'kode' => $request->kode,
-            'merek' => $request->merek,
-            'kondisi' => $request->kondisi,
-            'ket_produk' => $request->ket_produk,
-            'ket_kondisi' => $request->ket_kondisi,
-        ]); 
+        try {
+            $getCategorieName = DB::table('kategoris')->where('id', $request->kategori_id)->first()->kode;
+            $countProdSub = count(DB::table('produks')->where('kategori_id', $request->kategori_id)->where('sub_kategori_id', $request->sub_kategori_id)->get()) + 1;
+            $countProdKat = count(DB::table('produks')->where('kategori_id', $request->kategori_id)->get()) + 1;
+            // dd("$getCategorieName".count($countProd) + 1);
+
+            $produk->update([
+                'kategori_id' => $request->kategori_id,
+                'sub_kategori_id' => $request->sub_kategori_id,
+                'jurusan_id' => $request->jurusan_id,
+                'nama' => $request->nama,
+                'kode' => "$getCategorieName-$countProdKat-$countProdSub",
+                'merek' => $request->merek,
+                'kondisi' => $request->kondisi,
+                'ket_produk' => $request->ket_produk,
+                'ket_kondisi' => $request->ket_kondisi,
+            ]);
+
+            insertLog(Auth::user()->name . " Berhasil mengubah produk " . $request->nama);
+            return redirect()->route('produk.index')->with('msg_success', 'Berhasil mengubah produk');
+        } catch (\Throwable $th) {
+            return redirect()->route('produk.index')->with('msg_error', 'Gagal Diubah');
+        }
     }
 
     /**
@@ -115,8 +142,16 @@ class ProdukController extends Controller
      */
     public function destroy(Produk $produk)
     {
+        if (!$produk) {
+            return response()->json([
+                'message' => 'The data wanna be delete Not Found!'
+            ], 400);
+        }
+
         $produk->delete();
 
-        return back();
+        insertLog(Auth::user()->name . ' Berhasil menghapus produk ' . $produk->nama);
+
+        return redirect()->route('produk.index')->with('msg_success', 'Berhasil menghapus produk');
     }
 }
